requirements.txt
Django==1.11.4
django-cors-headers==2.1.0
django-jet==1.0.6
django-propeller==1.4.1
djangorestframework==3.6.4
gunicorn==19.7.1
psycopg2==2.7.3
pytz==2017.2
Unidecode==0.4.21
